+++ hyperion/importers/__init__.py
@@ -1,1 +1,2 @@
 from .orion import parse_orion
+from ._discretize_sph import _discretize_sph_func as discretize_sph